requirements_prod.txt
Django==2.2.4
django-filter==2.2.0
django-guardian==2.0.0
djangorestframework==3.10.2
Markdown==3.1.1
psycopg2-binary==2.8.3
Pygments==2.4.2
pytz==2019.2
sqlparse==0.3.0
